Added condition to use different auth guard depending on environment
The intention here is to use the ATS and HRIS guards only in production. And use the default Auth guard from angular when in development.
reason being so developers can navigate with the url and refresh without being logged out. But in production users are not permitted to do so and get logged out by the hris/ats guards